drawing logic extracted out of photoimport
